Exception handling for PUT/DELETE operations

+++ src/main/java/com/myproject/dscommerce/services/ProductService.java
@@ -1,15 +1,20 @@
 package com.myproject.dscommerce.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.myproject.dscommerce.dto.ProductDTO;
 import com.myproject.dscommerce.entities.Product;
 import com.myproject.dscommerce.repositories.ProductRepository;
+import com.myproject.dscommerce.services.exceptions.DataBaseException;
 import com.myproject.dscommerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -45,20 +50,33 @@ public class ProductService {
 	
 	@Transactional
 	public ProductDTO update(Long id, ProductDTO dto) {
-		
-		Product entity = repository.getReferenceById(id);
-		
-		copyDtoToEntity(dto, entity);
-		
-		entity = repository.save(entity);
-		
-		return new ProductDTO(entity);
+		try {
+			Product entity = repository.getReferenceById(id);
+
+			copyDtoToEntity(dto, entity);
+
+			entity = repository.save(entity);
+
+			return new ProductDTO(entity);
+		} 
+		catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+
 	}
 	
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
+		if(!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		try {
+			repository.deleteById(id);
+		}
+		catch(DataIntegrityViolationException e) {
+			throw new DataBaseException("Falha de integridade referencial");
+		}
 		
-		repository.deleteById(id);
 		
 	}
